+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changelog
 
 # 5.2.0 - 2026-08-10
-- Added an "Organisation block list" setting to the UI (General Settings) to manage the organisations blocked by "Block tracking requests from the cloud". Existing `block_geoip_organisations` config values are migrated to the new `organisation_block_list` system setting and removed from the config file; a removed or emptied config list results in the default block list applying again
+- Added an "Organisation block list" setting to the UI (General Settings, shown while "Block tracking requests from the cloud" is enabled) to manage the organisations blocked by that feature. Existing `block_geoip_organisations` config values are migrated to the new `organisation_block_list` system setting and removed from the config file; an emptied config list is migrated as an empty list, keeping organisation blocking disabled, while a removed config key results in the default block list applying again
 - The `trackingspamprevention:block-geo-ip-organisation` command now saves the organisation to the new system setting instead of the config file
 
 # 5.1.0 - 2026-08-03